Created AST Nodes and base for parser
I have added nodes which will be used to build the AST by the parser.
Also created a base for the parser along with its State manager. Lastly,
enabled copying and moving of the objects to prevent unnecessary
copying and invalid copies since the State managers for both the lexer
and parser use references which c++ does not shallowly copys.