WIP: Cleaned a bit and started expr parser
While starting the parser, I met a buttload of errors relating to
construction of nodes in which I've fixed by autogenerating the
constructors for the nodes, for now, only the expression nodes are fully
baked, the statements, partially. I'm using pratts parsing algorithm to
parse expressions which have precidences that are currently list in no
order, till later.