WIP: Fixed precedence and started statement parsing
The operator precedence is a bit wonky but made it the best I could
which marks the end of expression parsing using pratts algoritm. As for
statement parsing, a bunch of design issues have emerged one being on
error reporting which I find remarkably insufficient since the parser::error_t
members hold so little context which bloat if context is added, a
possible solution is delegating AST validation to the analyzer which
holds all info it needs to report any invalid syntax and others like
wrong use of break, continue and assignment to unassignable expressions.
This seems perfect, revision is under way.