Added statement parsing and node visit methods
As discussed regarding the design of the parser, the implemented one
takes that approach home. The parser is meant to simly build as much
nodes as it can, any encounter with unexpected tokens will not be errors
but will make the current construct lack some sections. These lacking
sections will be reported by the analyzer since it understands context
which is the main advantage. The parser will report basic errors such as
[unterminated,dangling] [braces,parens,brackets] and missing end of statement
token. This makes parsing very easy to make and maintain while the
analyzer will make sure the constructed nodes are complete else report
errors.