Get Node type for semantic analysis
This is a crucial piece of the analyzer: a visitor to get the type of
node at runtime. This should help walk through syntax fast and only
allow certain nodes in certain contexts.

Adding this needed a rewrite of the visitor classes which just never
worked. Unfortunately, the Visitor has to be one very big class, but
through preprocessor macros, we can override unnneded methods instantly.

Also, a bunch of errors are scattered every where but amazingly, the
fixes are easy to patch in thanks to the design.