Fixed eos granularity and expression statement
The original design to see semicolons and newlines as the same token,
Eos, did not age well, since newlines are like weak eos and semicolons
are strong. A semicolon can be seen by the program author while a
newline cannot. Some statements are allowed to span multiple lines which
makes newline invisible in this scope but semicolons mark eos with no
question. By sepparating both we get better control on which statements
can span multiple lines and which cannot, this is perfect.

Part of the language syntax is that a closing brace can act as eos for
the last statement in the block (it is also the eos marker for the block
whose start if the opening brace), therefore, a block with only an
expression statement cannot be differentiated fron a set with one item
by the analyzer, this calls for a wrapper on expressions making them
statements and not expressions directly.