Fixed precedence and assignment statement
The `as` keyword had a lower precedence compared to comma which means a
list of comma separated `as` expressions will not be parsed as a list of
aliases but a weird tree of alias to commas with other aliases as left
expression, just wrong! By raising the precedence, the expected is
achieved.

The parse_statement function gets stuck in an infinite loop if the
current token is not recognized as part of an expression which returns
an empty ExprStatement node with the original token not consumed and
leads to another iteration, checking if nullptr is returned from
assign_stmt function makes the token be wrapped in Statement node which
signifies an error to be reported later